pyproject.toml
[project]
name = "tunebot"
requires-python = ">=3.9"
version = "1.0.0"
dependencies = [
    "discord.py==2.3.2",
    "lavalink==5.9.0",
    "sqlalchemy==2.0.23",
    "asyncpg==0.29.0"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
